+++ source/city/bank/interfaces/BankTeller.java
@@ -4,7 +4,7 @@ import city.bank.BankCustomerRole;
 
 public interface BankTeller {
 	public void msgIAmHere(BankCustomer c);
-	public void msgHereIsMyRequest(BankCustomer c, String request, int amount);
+	public void msgHereIsMyRequest(BankCustomer c, String request, double amount);
 	public void msgLeavingBank(BankCustomer c);
 	public boolean isOccupied();
 	public void setOccupied(boolean b);
